Move `Not` impl to `operator` module.
